+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/base/database/tables/PostsMessagesTable.kt
@@ -2,9 +2,9 @@ package com.github.insanusmokrassar.AutoPostTelegramBot.base.database.tables
 
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.models.PostMessage
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.NewDefaultCoroutineScope
+import com.github.insanusmokrassar.AutoPostTelegramBot.utils.UnlimitedBroadcastChannel
 import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
 import kotlinx.coroutines.channels.BroadcastChannel
-import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.launch
 import org.jetbrains.exposed.sql.*
 import org.jetbrains.exposed.sql.transactions.transaction
@@ -14,12 +14,12 @@ typealias PostIdToMessagesIds = Pair<Int, Collection<MessageIdentifier>>
 val PostsMessagesTableScope = NewDefaultCoroutineScope()
 
 object PostsMessagesTable : Table() {
-    val newMessagesOfPost = BroadcastChannel<PostIdToMessagesIds>(Channel.CONFLATED)
+    val newMessagesOfPost: BroadcastChannel<PostIdToMessagesIds> = UnlimitedBroadcastChannel()
 
     @Deprecated("This channel is not determine post id", ReplaceWith("removedMessageOfPost"))
-    val removeMessageOfPost = BroadcastChannel<MessageIdentifier>(Channel.CONFLATED)
-    val removedMessagesOfPost = BroadcastChannel<PostIdToMessagesIds>(Channel.CONFLATED)
-    val removedMessageOfPost = BroadcastChannel<PostIdMessageId>(Channel.CONFLATED)
+    val removeMessageOfPost: BroadcastChannel<MessageIdentifier> = UnlimitedBroadcastChannel()
+    val removedMessagesOfPost: BroadcastChannel<PostIdToMessagesIds> = UnlimitedBroadcastChannel()
+    val removedMessageOfPost: BroadcastChannel<PostIdMessageId> = UnlimitedBroadcastChannel()
 
     private val messageId = long("messageId").primaryKey()
     private val mediaGroupId = text("mediaGroupId").nullable()