+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/base/BasePlugin.kt
@@ -36,7 +36,6 @@ class BasePlugin : Plugin {
         val botWR = WeakReference(executor)
 
         deletePost = DeletePost(
-            baseConfig.logsChatId,
             botWR
         )
         startPost = StartPost()