+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/rating/database/PostsLikesTable.kt
@@ -3,10 +3,10 @@ package com.github.insanusmokrassar.AutoPostTelegramBot.plugins.rating.database
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.database.tables.PostsTable
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.commonLogger
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.NewDefaultCoroutineScope
+import com.github.insanusmokrassar.AutoPostTelegramBot.utils.UnlimitedBroadcastChannel
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.subscribe
 import com.github.insanusmokrassar.TelegramBotAPI.types.ChatId
 import kotlinx.coroutines.channels.BroadcastChannel
-import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.launch
 import org.h2.jdbc.JdbcSQLException
 import org.jetbrains.exposed.sql.*
@@ -21,9 +21,9 @@ private const val resultColumnName = "result"
 private val PostsLikesTableScope = NewDefaultCoroutineScope()
 
 class PostsLikesTable : Table() {
-    val likesChannel = BroadcastChannel<PostIdUserId>(Channel.CONFLATED)
-    val dislikesChannel = BroadcastChannel<PostIdUserId>(Channel.CONFLATED)
-    val ratingsChannel = BroadcastChannel<PostIdRatingPair>(Channel.CONFLATED)
+    val likesChannel: BroadcastChannel<PostIdUserId> = UnlimitedBroadcastChannel()
+    val dislikesChannel: BroadcastChannel<PostIdUserId> = UnlimitedBroadcastChannel()
+    val ratingsChannel: BroadcastChannel<PostIdRatingPair> = UnlimitedBroadcastChannel()
 
     private val userId = long("userId").primaryKey()
     private val postId = integer("postId").primaryKey()
@@ -157,10 +157,9 @@ class PostsLikesTable : Table() {
     private fun userLike(userId: ChatId, postId: Int, like: Boolean) {
         val chooser = createChooser(postId, userId)
         transaction {
-            val record = select {
+            select {
                 chooser
-            }.firstOrNull()
-            record ?.let {
+            }.firstOrNull() ?.also {
                 if (it[this@PostsLikesTable.like] == like) {
                     deleteWhere { chooser }
                 } else {
@@ -168,18 +167,16 @@ class PostsLikesTable : Table() {
                         {
                             chooser
                         }
-                    ) {
-                        it[this@PostsLikesTable.like] = like
+                    ) { updateStatement ->
+                        updateStatement[this@PostsLikesTable.like] = like
                     }
                 }
-            } ?:let {
-                addUser(userId, postId, like)
-            }
-            PostsLikesTableScope.launch {
-                ratingsChannel.send(
-                    PostIdRatingPair(postId, getPostRating(postId))
-                )
-            }
+            } ?: addUser(userId, postId, like)
+        }
+        PostsLikesTableScope.launch {
+            ratingsChannel.send(
+                PostIdRatingPair(postId, getPostRating(postId))
+            )
         }
     }