+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/utils/extensions/Message.kt
@@ -9,13 +9,14 @@ private class MessageEntityDeformatter(
     private val messageEntity: MessageEntity? = null
 ) {
     val deformatted: String by lazy {
-        messageEntity?.asPreformattedPart(
-            sourceText
-        ) ?: sourceText.substring(
+        messageEntity ?.asMarkdownSource ?: sourceText.substring(
             range
         ).toMarkdown()
     }
 }
 
-@Deprecated("Now all this functionality is built-in Telegram Bot API")
+@Deprecated(
+    "Now all this functionality is built-in Telegram Bot API",
+    ReplaceWith("asMarkdownSource")
+)
 private fun MessageEntity.asPreformattedPart(stringToModify: String): String = asMarkdownSource